pkg/aflow: extract file-based code search tools
Extract DirIndex and ReadFile logic from pkg/codesearch into standalone
functions so they can be run without a built compilation database (*Index).

Move the aflow wrappers for these tools into a new filesystem.go file
and export them as FilesystemTools. This allows workflows like patch
triage to explore directories and read files without the overhead of
generating a full clang index.
